refactor(stores): decompose useCampaignRosterStore.ts into 3 modules
The campaign roster store exceeded the 600-LOC critical threshold
(618 LOC). Split it into three co-located sibling modules with zero
behavior change — the Zustand store shape, action signatures, persist
config, and all public exports are byte-for-byte preserved.

- campaignRosterStore.types.ts (173 LOC) — the full type surface:
  UnitReadiness, ICampaignMissionRecord, IUnitDamageState,
  CampaignRosterState, CampaignRosterActions, CampaignRosterStore.
- campaignRosterStore.helpers.ts (260 LOC) — pure helper functions
  with no store coupling: generateId, applyDamageToCanonicalState,
  plus newly extracted derivations (deriveReadinessFromCarryForward,
  refreshUnitsReadiness, buildNextCombatStateMap) lifted out of the
  applyDamageCarryForward action body.
- useCampaignRosterStore.ts (618 -> 318 LOC) — now only the Zustand
  store definition. The public type names are re-exported from this
  module so every existing importer resolves them unchanged.

Verified: tsc --noEmit clean, oxlint src/stores 0 warnings/errors,
627 store + roster integration tests pass.
